# servers/customer-services/docs/api-documentation.yaml
openapi: 3.1.0
info:
  title: InsightSerenity Customer Services API
  version: 1.0.0
  description: |
    # Customer Services API Documentation

    ## Overview
    The Customer Services API provides multi-tenant business services including:
    - User Management
    - Client Management
    - Project Management
    - Recruitment Services
    - Organization Management

    ## Authentication
    All endpoints require JWT Bearer token authentication.

    ## Multi-tenancy
    All requests must include the `X-Tenant-ID` header.
servers:
  - url: http://localhost:3001
    description: Development Server
  - url: https://api.insightserenity.com
    description: Production Server
security:
  - bearerAuth: []
...
paths:
...
  /api/users/{userId}:
    get:
      tags: [Users]
      summary: Get user by ID
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: User retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'

    put:
      tags: [Users]
      summary: Update user
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        200:
          description: User updated successfully

    delete:
      tags: [Users]
      summary: Delete user
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: User deleted successfully
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantIdHeader:
      name: X-Tenant-ID
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Tenant identifier
...
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        role:
          type: string
          enum: [super_admin, admin, manager, employee, client, consultant]
        status:
          type: string
          enum: [active, inactive, pending, suspended, deleted]
        profile:
          type: object
          properties:
            displayName:
              type: string
            avatar:
              type: object
              properties:
                url:
                  type: string
            bio:
              type: string
            phoneNumbers:
              type: array
              items:
                type: object
                properties:
                  number:
                    type: string
                  type:
                    type: string
                    enum: [mobile, home, work]
        professional:
          type: object
          properties:
            jobTitle:
              type: string
            department:
              type: string
            employeeId:
              type: string
        emailVerified:
          type: boolean
        twoFactorEnabled:
          type: boolean
        lastLogin:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        profile:
          type: object
        professional:
          type: object
        status:
          type: string
          enum: [active, inactive, suspended]
...